Added new methods to support custom formatting of floats and doubles:
-setFloatFormat: (char *) format;

If this method is not invoked - the formatting for display defaults
to that set in the global probeLibrary instance.
If there was no display defaults set in the probeLibrary - they default
to the standard "%g" formatting which preserves backwards compatibility.

-(char *) probeAsString: (id) anObject Buffer: (char *) buf
      withFullPrecision: (int) precision;

This method added to support fuller precision probing for saving via
ObjectSaver - not needed by a user.